refactor: 예외 위치 수정
