+++ src/main/java/roomescape/global/exception/InvalidTokenException.java
@@ -1,12 +1,11 @@
-package roomescape.member;
+package roomescape.global.exception;
 
 import org.springframework.http.HttpStatus;
 
-public class InvalidMemberException extends RuntimeException {
-
+public class InvalidTokenException extends RuntimeException {
     private final HttpStatus statusCode;
 
-    public InvalidMemberException(String message, HttpStatus statusCode) {
+    public InvalidTokenException(HttpStatus statusCode, String message) {
         super(message);
         this.statusCode = statusCode;
     }
