+++ src/main/java/roomescape/reservation/application/service/ReservationService.java
@@ -9,8 +9,8 @@ import roomescape.admin.presentation.dto.AdminReservationRequest;
 import roomescape.global.exception.DuplicateReservationException;
 import roomescape.global.exception.GetThemeException;
 import roomescape.global.exception.GetTimeException;
+import roomescape.global.exception.InvalidMemberException;
 import roomescape.global.exception.PastTimeException;
-import roomescape.member.InvalidMemberException;
 import roomescape.member.application.repository.MemberRepository;
 import roomescape.member.domain.Member;
 import roomescape.reservation.application.dto.CreateReservationRequest;
